requirements.txt
black
ruff
pytest
pyyaml
jsonschema
